Update handling of audio files
