+++ src/data/animalList.js
@@ -128,7 +128,7 @@ const animalList = [
             letter: 'x',
             animal: 'xrayfish',
             image: 'img_x.svg',
-            sound: 'x-ray-fish.mp3',
+            sound: 'xrayfish.mp3',
         },
         {
             letter: 'c',
